Refactor SellAction component and remove TransactionAction
- Replaced TransactionAction with a new SellAction component that handles selling options.
- Implemented asset selection, price formatting, and validation for sell options.
- Added treasury balance display and error handling for missing treasury wallet.
- Created shared transaction helper functions for validation and price formatting.
- Removed the old TransactionAction component as it is no longer needed.
