--- src/components/modals/CreateProposalModal/MarketActions/TransactionAction.jsx
@@ -1,662 +1,0 @@
-import { useEffect, useMemo, useState, useCallback } from 'react';
-import { AlertCircle, Plus, Wallet, X, Loader2, CheckCircle } from 'lucide-react';
-
-import { LavaSteelInput } from '@/components/shared/LavaInput';
-import { LavaMultiSelect, LavaSteelSelect } from '@/components/shared/LavaSelect';
-import { useVaultAssetsForProposalByType, useAssetMetadata } from '@/services/api/queries';
-import { LavaIntervalPicker } from '@/components/shared/LavaIntervalPicker';
-import { LavaCheckbox } from '@/components/shared/LavaCheckbox';
-import { transactionOptionSchema, buyOptionSchema } from '@/components/vaults/constants/proposal.constants.js';
-import { IS_MAINNET } from '@/utils/networkValidation.ts';
-
-const methodOptions = [
-  { value: 'N/A', label: 'Time Limit' },
-  { value: 'GTC', label: 'GTC' },
-];
-
-const sellTypeOptions = [
-  { value: 'List', label: 'List' },
-  { value: 'Market', label: 'Market' },
-];
-
-const buyTypeOptions = [
-  { value: 'Offer', label: 'Offer' },
-  { value: 'Buy', label: 'Buy' },
-];
-
-const validateOptions = (options, isBuyType = false) => {
-  if (options.length === 0) return false;
-
-  const schema = isBuyType ? buyOptionSchema : transactionOptionSchema;
-
-  try {
-    options.forEach(option => {
-      schema.validateSync(option);
-    });
-    return true;
-  } catch {
-    return false;
-  }
-};
-
-const validateOption = (option, isBuyType = false) => {
-  if (!option) return false;
-
-  const schema = isBuyType ? buyOptionSchema : transactionOptionSchema;
-
-  try {
-    schema.validateSync(option);
-    return true;
-  } catch {
-    return false;
-  }
-};
-
-// Component to handle asset metadata fetching for a single buy option
-const BuyOptionMetadataHandler = ({ option, onMetadataChange }) => {
-  const [debouncedUnit, setDebouncedUnit] = useState('');
-
-  // Validate unit format: more than 56 hex characters (policy + asset name required)
-  const isValidUnit = Boolean(option.assetId && /^[a-fA-F0-9]{57,}$/.test(option.assetId));
-
-  // Clear display name immediately when unit changes
-  useEffect(() => {
-    onMetadataChange(option.id, {
-      displayName: '',
-      unit: option.assetId || '',
-      isLoading: false,
-      error: null,
-    });
-  }, [option.assetId, option.id, onMetadataChange]);
-
-  // Debounce the unit value
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setDebouncedUnit(option.assetId || '');
-    }, 300);
-
-    return () => clearTimeout(timer);
-  }, [option.assetId]);
-
-  // Fetch metadata when debounced unit is valid
-  const { data, isLoading, error } = useAssetMetadata(debouncedUnit, isValidUnit);
-
-  // Update parent component when metadata changes
-  useEffect(() => {
-    // Only update if the response is for the current unit (prevent stale data)
-    const currentUnit = option.assetId || '';
-
-    if (data?.data && debouncedUnit === currentUnit) {
-      onMetadataChange(option.id, {
-        displayName: data.data.displayName,
-        unit: currentUnit,
-        isLoading: false,
-        error: null,
-      });
-    } else if (error && debouncedUnit === currentUnit) {
-      onMetadataChange(option.id, {
-        displayName: '',
-        unit: currentUnit,
-        isLoading: false,
-        error: error.message || 'Failed to fetch asset metadata',
-      });
-    } else if (isLoading && debouncedUnit === currentUnit) {
-      onMetadataChange(option.id, {
-        displayName: '',
-        unit: currentUnit,
-        isLoading: true,
-        error: null,
-      });
-    }
-  }, [data, isLoading, error, option.id, option.assetId, debouncedUnit, onMetadataChange]);
-
-  return null; // This component doesn't render anything
-};
-
-const TransactionAction = ({ vaultId, onDataChange, error, execType, title = 'Transaction Options' }) => {
-  const [options, setOptions] = useState([]);
-  const [assetOptions, setAssetOptions] = useState([]);
-  const [abstain, setAbstain] = useState(false);
-  const [assetMetadataMap, setAssetMetadataMap] = useState({}); // Map of optionId -> metadata
-
-  const isBuyType = execType === 'BUY';
-  const getExecValue = sellType => {
-    if (!isBuyType) return execType;
-    return sellType === 'Offer' ? 'OFFER' : 'BUY';
-  };
-
-  // Handle metadata changes for buy options
-  const handleMetadataChange = useCallback((optionId, metadata) => {
-    setAssetMetadataMap(prev => ({
-      ...prev,
-      [optionId]: metadata,
-    }));
-
-    // Update the option with assetName (for validation) and displayName (for historical reference)
-    setOptions(prevOptions =>
-      prevOptions.map(opt =>
-        opt.id === optionId
-          ? {
-              ...opt,
-              assetName: metadata.displayName || '',
-              displayName: metadata.displayName || '',
-            }
-          : opt
-      )
-    );
-  }, []);
-
-  const { data, isLoading } = useVaultAssetsForProposalByType(vaultId, 'buy-sell');
-
-  const assetsData = data?.data.assets;
-  const treasuryInfo = data?.data?.treasuryWalletBalance;
-  const treasuryBalance = treasuryInfo?.lovelace / 1000000 || 0;
-
-  const remainingAssets = useMemo(() => {
-    if (!assetsData) return [];
-    const usedAssetNames = options.map(option => option.assetName).filter(Boolean);
-    return assetsData.filter(asset => !usedAssetNames.includes(asset.name));
-  }, [assetsData, options]);
-
-  const selectedAssets = useMemo(() => {
-    return options.map(opt => (isBuyType ? opt.assetId : opt.assetName)).filter(Boolean);
-  }, [options, isBuyType]);
-
-  useEffect(() => {
-    if (assetsData && !isLoading) {
-      const formattedAssets = assetsData.map(asset => ({
-        value: asset.name,
-        label: asset.name,
-        id: asset.id,
-      }));
-
-      setAssetOptions(formattedAssets);
-    }
-  }, [assetsData, isLoading]);
-
-  useEffect(() => {
-    onDataChange({
-      buyingSellingOptions: options,
-      abstain,
-      isValid: validateOptions(options, isBuyType),
-    });
-  }, [options, onDataChange, abstain, isBuyType]);
-
-  const handleOptionChange = (id, field, value) => {
-    if (field === 'assetName') {
-      if (isBuyType) {
-        setOptions(
-          options.map(option =>
-            option.id === id ? { ...option, assetName: value, exec: getExecValue(option.sellType) } : option
-          )
-        );
-      } else {
-        const selectedAsset = assetOptions.find(option => option.value === value);
-        setOptions(
-          options.map(option =>
-            option.id === id
-              ? {
-                  ...option,
-                  [field]: value,
-                  assetId: selectedAsset?.id || null,
-                  exec: getExecValue(option.sellType),
-                }
-              : option
-          )
-        );
-      }
-    } else if (field === 'assetId') {
-      setOptions(
-        options.map(option =>
-          option.id === id
-            ? {
-                ...option,
-                assetId: value,
-                exec: getExecValue(option.sellType),
-              }
-            : option
-        )
-      );
-    } else if (field === 'sellType') {
-      setOptions(
-        options.map(option =>
-          option.id === id ? { ...option, [field]: value, price: '', exec: getExecValue(value) } : option
-        )
-      );
-    } else if (field === 'method') {
-      setOptions(
-        options.map(option =>
-          option.id === id ? { ...option, [field]: value, duration: '', exec: getExecValue(option.sellType) } : option
-        )
-      );
-    } else {
-      setOptions(
-        options.map(option =>
-          option.id === id ? { ...option, [field]: value, exec: getExecValue(option.sellType) } : option
-        )
-      );
-    }
-  };
-
-  const setFTMax = (id, checked) => {
-    const option = options.find(o => o.id === id);
-    const asset = option && assetsData?.data ? assetsData.data.find(a => a.name === option.assetName) : null;
-
-    setOptions(prev =>
-      prev.map(o =>
-        o.id === id
-          ? {
-              ...o,
-              isMax: checked,
-              quantity: checked ? String(asset?.quantity ?? '') : '',
-            }
-          : o
-      )
-    );
-  };
-
-  const getAvailableAmount = id => {
-    const option = options.find(o => o.id === id);
-    if (option && assetsData?.data) {
-      const asset = assetsData.data.find(a => a.name === option.assetName);
-      return asset ? (+asset.quantity).toFixed(0) : 0;
-    }
-
-    return 0;
-  };
-
-  const handleAmountChange = (id, field, value) => {
-    const parts = value.split('.');
-    if (parts.length > 2) value = parts[0] + '.' + parts.slice(1).join('');
-
-    if (value.includes('.')) {
-      const [int, dec] = value.split('.');
-      value = int + '.' + dec.slice(0, 2);
-    }
-
-    if (Number(value) <= 0) {
-      setOptions(options.map(option => (option.id === id ? { ...option, [field]: '' } : option)));
-      return;
-    }
-
-    setOptions(options.map(option => (option.id === id ? { ...option, [field]: value } : option)));
-  };
-
-  const handleAddOption = selectedAssetValues => {
-    if (options.length >= 10) return;
-
-    if (selectedAssetValues && Array.isArray(selectedAssetValues)) {
-      const newOptions = selectedAssetValues.map(assetValue => {
-        const existingOption = options.find(opt =>
-          isBuyType ? opt.assetId === assetValue : opt.assetName === assetValue
-        );
-
-        if (existingOption) {
-          return existingOption;
-        }
-
-        const selectedAsset = assetOptions.find(opt => opt.value === assetValue);
-        return {
-          id: Date.now() + Math.random(),
-          assetName: selectedAsset?.label || '',
-          assetId: selectedAsset?.id || (isBuyType ? assetValue : null),
-          exec: getExecValue(isBuyType ? 'Buy' : ''),
-          quantity: '',
-          sellType: isBuyType ? 'Buy' : '',
-          duration: '',
-          isMax: false,
-          method: 'N/A',
-          market: 'WayUp',
-          price: '',
-        };
-      });
-
-      setOptions(newOptions);
-      return;
-    }
-
-    setOptions([
-      ...options,
-      {
-        id: Date.now(),
-        assetName: '',
-        assetId: isBuyType ? '' : null,
-        exec: getExecValue(isBuyType ? 'Buy' : ''),
-        quantity: '',
-        sellType: isBuyType ? 'Buy' : '',
-        duration: '',
-        isMax: false,
-        method: 'N/A',
-        market: 'WayUp',
-        price: '',
-      },
-    ]);
-  };
-
-  const handleRemoveOption = id => setOptions(options.filter(option => option.id !== id));
-
-  if (IS_MAINNET && !treasuryInfo) {
-    return (
-      <div className="bg-red-900/20 border border-red-500/30 rounded-lg p-4">
-        <div className="flex items-center gap-3">
-          <AlertCircle className="w-5 h-5 text-red-400" />
-          <div>
-            <p className="text-red-400 font-medium">No Treasury Wallet</p>
-            <p className="text-white/60 text-sm mt-1">
-              This vault does not have a treasury wallet configured. Distribution proposals require a treasury wallet
-              with ADA funds.
-            </p>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <>
-      <div className="bg-steel-800 rounded-lg p-4">
-        <div className="flex items-center gap-2 text-white/60 mb-2">
-          <Wallet className="w-4 h-4" />
-          <span className="text-sm">Treasury Balance</span>
-        </div>
-        <p className="text-2xl font-semibold text-white">{treasuryBalance.toLocaleString()} ADA</p>
-      </div>
-      <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center mb-4 gap-3">
-        <h3 className="text-lg font-medium">{title}</h3>
-        <div className="flex flex-col sm:flex-row gap-3 sm:w-auto w-[100%]">
-          <button
-            className="flex items-center justify-center gap-2 bg-steel-850 hover:bg-steel-850/70 text-white/60 px-4 py-2 rounded-lg transition-colors w-full border border-steel-750"
-            type="button"
-            disabled={options.length >= 10}
-            onClick={handleAddOption}
-          >
-            Add Action
-            <Plus className="h-4 w-4" />
-          </button>
-          {!isBuyType && (
-            <LavaMultiSelect
-              options={assetOptions.map(asset => ({
-                label: asset.label,
-                value: asset.value,
-              }))}
-              value={selectedAssets}
-              placeholder="Add Multiple NFTs"
-              onChange={handleAddOption}
-              className="min-w-[250px]"
-            />
-          )}
-        </div>
-      </div>
-
-      {options.length === 0 ? (
-        <>
-          {error && <p className="text-center text-red-600 font-bold py-8">Add actions!</p>}
-          {!error && <p className="text-center text-white/60 py-8">Start by clicking Add action</p>}
-        </>
-      ) : (
-        <div className="space-y-8">
-          {options.map((option, index) => {
-            const isOverLimit = parseFloat(option.quantity) > getAvailableAmount(option.id);
-            const metadata = assetMetadataMap[option.id] || {};
-
-            return (
-              <div key={option.id}>
-                {isBuyType && <BuyOptionMetadataHandler option={option} onMetadataChange={handleMetadataChange} />}
-                <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center mb-4 gap-2">
-                  <p className="font-medium">
-                    Option {index + 1}{' '}
-                    {error && !validateOption(option, isBuyType) && (
-                      <span className="text-red-600 ml-2">
-                        {isBuyType && option.assetId && option.assetId.length > 0 && option.assetId.length < 56
-                          ? 'Asset ID must be at least 56 characters!'
-                          : 'Fill in all inputs!'}
-                      </span>
-                    )}
-                  </p>
-                  <button
-                    className="bg-red-600/10 hover:bg-red-600/20 text-red-600 text-sm px-3 py-1 rounded-md flex items-center gap-1.5 transition-colors self-start sm:self-auto"
-                    type="button"
-                    onClick={() => handleRemoveOption(option.id)}
-                  >
-                    <X className="h-4 w-4" />
-                    Remove
-                  </button>
-                </div>
-                <div className="relative bg-steel-800 p-4 rounded-[10px]">
-                  {isBuyType ? (
-                    <>
-                      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4 mb-6">
-                        <div>
-                          <div className="flex items-center justify-between gap-2 mb-2">
-                            <p className="text-sm text-gray-400">Display Name:</p>
-                            {metadata.displayName && metadata.unit === option.assetId && option.assetId.length > 56 && (
-                              <a
-                                href={`https://www.wayup.io/collection/${option.assetId.substring(0, 56)}/asset/${option.assetId.substring(56)}`}
-                                target="_blank"
-                                rel="noopener noreferrer"
-                                className="text-xs text-orange-500 hover:text-orange-400 hover:underline"
-                              >
-                                WayUp
-                              </a>
-                            )}
-                          </div>
-                          {metadata.isLoading ? (
-                            <div className="flex items-center gap-2 px-3 py-2.5 bg-steel-900/50 border border-steel-700/50 rounded-lg">
-                              <Loader2 className="w-4 h-4 animate-spin text-orange-500 flex-shrink-0" />
-                              <span className="text-sm text-gray-400">Loading asset metadata...</span>
-                            </div>
-                          ) : metadata.displayName && metadata.unit === option.assetId ? (
-                            <div className="flex items-center gap-2 px-3 py-2.5 bg-green-900/20 border border-green-500/30 rounded-lg">
-                              <CheckCircle className="w-4 h-4 text-green-500 flex-shrink-0" />
-                              <span className="text-sm text-gray-200 break-words">{metadata.displayName}</span>
-                            </div>
-                          ) : metadata.error && metadata.unit === option.assetId ? (
-                            <div className="px-3 py-2.5 bg-red-900/20 border border-red-500/30 rounded-lg">
-                              <p className="text-xs text-red-400">{metadata.error}</p>
-                            </div>
-                          ) : (
-                            <div className="px-3 py-2.5 bg-steel-900/50 border border-steel-750 rounded-lg">
-                              <span className="text-sm text-gray-500 italic">Enter asset unit</span>
-                            </div>
-                          )}
-                        </div>
-                        <div>
-                          <p className="text-sm text-gray-400 mb-2">Exec Type</p>
-                          <LavaSteelSelect
-                            options={buyTypeOptions}
-                            placeholder="Select type"
-                            value={option.sellType ?? buyTypeOptions[1].value}
-                            onChange={value => handleOptionChange(option.id, 'sellType', value)}
-                          />
-                        </div>
-                        <div>
-                          <p className="text-sm text-gray-400 mb-2">Market</p>
-                          <LavaSteelSelect
-                            placeholder="Select market"
-                            value={option.market}
-                            onChange={value => handleOptionChange(option.id, 'market', value)}
-                            options={[{ value: 'WayUp', label: 'WayUp' }]}
-                          />
-                        </div>
-                        <div>
-                          <p className="text-sm text-gray-400 mb-2">
-                            {option.sellType === 'Offer' ? 'Offer Price (ADA)' : 'Max Buy Price (ADA)'}
-                          </p>
-                          <LavaSteelInput
-                            type="number"
-                            min={0}
-                            step={0.1}
-                            placeholder="0.00"
-                            value={option.price}
-                            onChange={value => handleAmountChange(option.id, 'price', value)}
-                            onIncrement={() => {
-                              const newValue = (parseFloat(option.price) || 0) + 0.1;
-                              handleAmountChange(option.id, 'price', newValue.toFixed(1));
-                            }}
-                            onDecrement={() => {
-                              const newValue = Math.max(0, (parseFloat(option.price) || 0) - 0.1);
-                              handleAmountChange(option.id, 'price', newValue.toFixed(1));
-                            }}
-                          />
-                        </div>
-                      </div>
-                      <div>
-                        <p className="text-sm text-gray-400 mb-2">Asset Unit:</p>
-                        <LavaSteelInput
-                          type="text"
-                          placeholder="Enter asset unit"
-                          value={option.assetId || ''}
-                          onChange={value => handleOptionChange(option.id, 'assetId', value)}
-                          className={
-                            option.assetId && option.assetId.length > 0 && option.assetId.length < 56
-                              ? '!border-red-500/60'
-                              : ''
-                          }
-                        />
-                        {option.assetId && option.assetId.length > 0 && option.assetId.length < 56 && (
-                          <p className="text-xs text-red-500 mt-1">Asset Unit must be at least 56 characters</p>
-                        )}
-                      </div>
-                    </>
-                  ) : (
-                    <>
-                      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-                        <div>
-                          <p className="text-sm text-gray-400 mb-2">Asset Name:</p>
-                          <LavaSteelSelect
-                            options={assetOptions.filter(
-                              opt => opt.value === option.assetName || remainingAssets.some(a => a.name === opt.value)
-                            )}
-                            placeholder={isLoading ? 'Loading assets...' : 'Select asset'}
-                            value={option.assetName}
-                            onChange={value => handleOptionChange(option.id, 'assetName', value)}
-                            isDisabled={isLoading}
-                          />
-                        </div>
-                        <div>
-                          <div className="flex justify-between gap-2 mb-1">
-                            <p className="text-sm text-gray-400">Quantity</p>
-                            <LavaCheckbox
-                              name={`max-${option.id}`}
-                              checked={option.isMax}
-                              label={`max: ${getAvailableAmount(option.id)}`}
-                              className="whitespace-nowrap"
-                              labelClassName="text-gray-400"
-                              onChange={e => setFTMax(option.id, e.target.checked)}
-                              disabled={!option.assetName}
-                            />
-                          </div>
-                          <LavaSteelInput
-                            type="number"
-                            min={0}
-                            max={getAvailableAmount(option.id)}
-                            placeholder="0.00"
-                            value={option.quantity}
-                            onChange={value => {
-                              const numValue = parseFloat(value);
-                              if (value === '' || (numValue >= 0 && numValue <= getAvailableAmount(option.id))) {
-                                handleOptionChange(option.id, 'quantity', value);
-                              }
-                            }}
-                            onIncrement={() => {
-                              const newValue = (parseFloat(option.quantity) || 0) + 1;
-                              if (newValue <= getAvailableAmount(option.id)) {
-                                handleOptionChange(option.id, 'quantity', newValue.toString());
-                              }
-                            }}
-                            onDecrement={() => {
-                              const newValue = Math.max(0, (parseFloat(option.quantity) || 0) - 1);
-                              handleOptionChange(option.id, 'quantity', newValue.toString());
-                            }}
-                            className={isOverLimit ? '!border-red-500/60' : ''}
-                          />
-                        </div>
-                        <div>
-                          <p className="text-sm text-gray-400 mb-2">Sell Type</p>
-                          <LavaSteelSelect
-                            options={sellTypeOptions}
-                            placeholder="Select type"
-                            value={option.sellType}
-                            onChange={value => handleOptionChange(option.id, 'sellType', value)}
-                          />
-                        </div>
-                      </div>
-                      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-4 mt-4">
-                        <div>
-                          <p className="text-sm text-gray-400 mb-2">Method</p>
-                          <LavaSteelSelect
-                            options={methodOptions}
-                            placeholder="Select method"
-                            value={option.method}
-                            onChange={value => handleOptionChange(option.id, 'method', value)}
-                          />
-                        </div>
-                        <div>
-                          <p className="text-sm text-gray-400 -mb-2">Duration</p>
-                          <LavaIntervalPicker
-                            placeholder="Select"
-                            value={option.duration}
-                            variant="steel"
-                            onChange={value => handleOptionChange(option.id, 'duration', value)}
-                            disabled={option.method === 'GTC'}
-                          />
-                        </div>
-                        <div>
-                          <p className="text-sm text-gray-400 mb-2">Market</p>
-                          <LavaSteelSelect
-                            placeholder="Select market"
-                            value={option.market}
-                            onChange={value => handleOptionChange(option.id, 'market', value)}
-                            options={[{ value: 'WayUp', label: 'WayUp' }]}
-                          />
-                        </div>
-                        <div>
-                          <p className="text-sm text-gray-400 mb-2">Price</p>
-                          <LavaSteelInput
-                            type="number"
-                            min={0}
-                            step={0.1}
-                            placeholder="0.00"
-                            value={option.price}
-                            onChange={value => handleAmountChange(option.id, 'price', value)}
-                            onIncrement={() => {
-                              const newValue = (parseFloat(option.price) || 0) + 0.1;
-                              handleAmountChange(option.id, 'price', newValue.toFixed(1));
-                            }}
-                            onDecrement={() => {
-                              const newValue = Math.max(0, (parseFloat(option.price) || 0) - 0.1);
-                              handleAmountChange(option.id, 'price', newValue.toFixed(1));
-                            }}
-                            disabled={option.sellType === 'Market'}
-                          />
-                        </div>
-                      </div>
-                    </>
-                  )}
-                </div>
-              </div>
-            );
-          })}
-          {abstain && (
-            <div>
-              <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center mb-4 gap-2">
-                <p className="font-medium">Option {options.length + 1}</p>
-              </div>
-              <div className="relative bg-steel-800 p-4 rounded-[10px]">
-                <p className="font-medium">Do nothing</p>
-              </div>
-            </div>
-          )}
-        </div>
-      )}
-      <LavaCheckbox
-        checked={abstain}
-        label="Abstain"
-        labelClassName="text-[20px]"
-        onChange={e => setAbstain(e.target.checked)}
-      />
-    </>
-  );
-};
-
-export default TransactionAction;
