+++ src/components/vaults/constants/proposal.constants.js
@@ -20,7 +20,6 @@ export const buyOptionSchema = yup.object({
 
 export const transactionOptionSchema = yup.object({
   assetName: yup.string().required(),
-  quantity: yup.string().required(),
   duration: yup.string().when('method', {
     is: 'GTC',
     then: schema => schema.notRequired().nullable(),
@@ -35,6 +34,5 @@ export const transactionOptionSchema = yup.object({
   sellType: yup.string().required(),
   exec: yup.string().required(),
   id: yup.number().required(),
-  isMax: yup.boolean().required(),
   market: yup.string().required(),
 });
